SqlCaseUpdateStrategy.fetch_case_transaction_forms
